[feat] 잘못 올린 java 파일 삭제 및 application.yml 원상복구.
