--- java/com/mc/config/SwaggerConfig.java
@@ -1,23 +1,0 @@
-package com.mc.config;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-
-    @Bean
-    public OpenAPI openAPI() {
-        return new OpenAPI()
-                .info(apiInfo());
-    }
-
-    private Info apiInfo() {
-        return new Info()
-                .title("API 명세서")
-                .description("게스트 웹 서비스 API 명세서")
-                .version("v1");
-    }
-}
