--- java/com/mc/config/WebMvcConfig.java
@@ -1,21 +1,0 @@
-package com.mc.config;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Slf4j
-@Configuration
-public class WebMvcConfig implements WebMvcConfigurer {
-
-    @Value("${app.dir.imgdir}")
-    String imgdir;
-
-    @Override
-    public void addResourceHandlers(ResourceHandlerRegistry registry) {
-        registry.addResourceHandler("/imgRepo/**").addResourceLocations(imgdir);
-    }
-
-}
